Flatten nested package into a single directory
