--- firecrown/models/cluster/integrator/__init__.py
@@ -1,13 +1,0 @@
-"""Module for cluster integrator classes."""
-
-from firecrown.models.cluster.integrator._integrator import Integrator
-from firecrown.models.cluster.integrator._numcosmo_integrator import (
-    NumCosmoIntegrator,
-)
-from firecrown.models.cluster.integrator._scipy_integrator import ScipyIntegrator
-
-__all__ = [
-    "Integrator",
-    "NumCosmoIntegrator",
-    "ScipyIntegrator",
-]
