--- firecrown/models/cluster/recipes/__init__.py
@@ -1,15 +1,0 @@
-"""Module for cluster recipe classes."""
-
-from firecrown.models.cluster.recipes._cluster_recipe import ClusterRecipe
-from firecrown.models.cluster.recipes._murata_binned_spec_z import (
-    MurataBinnedSpecZRecipe,
-)
-from firecrown.models.cluster.recipes._murata_binned_spec_z_deltasigma import (
-    MurataBinnedSpecZDeltaSigmaRecipe,
-)
-
-__all__ = [
-    "ClusterRecipe",
-    "MurataBinnedSpecZRecipe",
-    "MurataBinnedSpecZDeltaSigmaRecipe",
-]
