+++ tests/cluster_recipes/test_murata_binned_spec_z_delta_sigma.py
@@ -13,8 +13,8 @@ from firecrown.models.cluster import (
     NDimensionalBin,
     SpectroscopicRedshift,
 )
-from firecrown.models.cluster.integrator import NumCosmoIntegrator
-from firecrown.models.cluster.recipes import (
+from firecrown.models.cluster import NumCosmoIntegrator
+from firecrown.models.cluster import (
     ClusterRecipe,
     MurataBinnedSpecZDeltaSigmaRecipe,
     MurataBinnedSpecZRecipe,
